+++ spec/integration/schema/inferrer/postgres_spec.rb
@@ -32,6 +32,9 @@ RSpec.describe 'ROM::SQL::Schema::PostgresInferrer', :postgres do
       lseg :lseg
       polygon :polygon
       path :path
+      timestamp :created_at
+      column :datetime, "timestamp(0) without time zone"
+      column :datetime_tz, "timestamp(0) with time zone"
     end
   end
 
@@ -113,7 +116,10 @@ RSpec.describe 'ROM::SQL::Schema::PostgresInferrer', :postgres do
           name: :path,
           source: source,
           read: ROM::SQL::Types::PG::PathTR.optional
-        )
+        ),
+        created_at: ROM::SQL::Types::Time.optional.meta(name: :created_at, source: source),
+        datetime: ROM::SQL::Types::Time.optional.meta(name: :datetime, source: source),
+        datetime_tz: ROM::SQL::Types::Time.optional.meta(name: :datetime_tz, source: source)
       )
     end
   end