+++ spec/spec_helper.rb
@@ -16,7 +16,8 @@ require 'tempfile'
 
 begin
   require 'byebug'
-rescue LoadError # rubocop:disable Lint/HandleExceptions
+rescue LoadError
+  require 'pry'
 end
 
 LOGGER = Logger.new(File.open('./log/test.log', 'a'))
@@ -40,10 +41,13 @@ PG_LTE_95 = ENV.fetch('PG_LTE_95', 'true') == 'true'
 
 SPEC_ROOT = root = Pathname(__FILE__).dirname
 
-# Redirect inferrer warnings to a log file
-$stderr.reopen(SPEC_ROOT.join('../log/warnings.log'), 'w')
-
 TMP_PATH = root.join('../tmp')
+
+class ROM::SQL::Schema::Inferrer
+  def self.on_error(*)
+    # quiet in specs
+  end
+end
 
 Dir[root.join('shared/**/*')].each { |f| require f }
 Dir[root.join('support/**/*')].each { |f| require f }