+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Support for inferring the PostgreSQL `hstore` data type (flash-gordon)
 * Support for the rest of geometric PostgreSQL data types (box, lseg, polygon, etc.) (Morozzzko)
 * Added inferring for timestamp types with specified precision (flash-gordon)
+* Added `ROM::SQL::Attribute#in` to support range checks in conditions (flash-gordon)
 
 ### Fixed